+++ ChangeLog.md
@@ -37,7 +37,8 @@ ChangeLog
 * #567: Doing freebusy requests on accounts that had 0 calendars would throw
   a `E_NOTICE`.
 * #572: `HEAD` requests trigger a PHP warning.
-
+* #579: Browser plugin can throw exception for a few resourcetypes that didn't
+  have an icon defined.
 
 
 2.1.1 (2014-11-22)